old-tests/submission/W3C/domhtml/HTMLOListElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OLIST</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-ol-element.html#dom-ol-start" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<ol compact="compact" start="1" type="1">
<li>EMP0001
<ul>
<li>Margaret Martin
<dl>
<dd>Accountant</dd>
<dd>56,000</dd>
</dl>
</li>
</ul>
</li>
<li>EMP0002
<ul>
<li>Martha Raynolds
<dl>
<dd>Secretary</dd>
<dd>35,000</dd>
</dl>
</li>
</ul>
</li>
</ol>



<script>
var domtest = async_test("The start attribute specifies the starting sequence number.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vstart;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("ol");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vstart = testNode.start;

      assert_equals(vstart,1);
       
    domtest.done();
}));

</script>
<p>The start attribute specifies the starting sequence number.        

    Retrieve the start attribute and examine its value.</p>
<div id="log"></div>
</body></html>
